fix/4 eventpage
